requirements.txt
# VoiceBot Requirements
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2

# ML / NLP
torch==2.4.1
transformers==4.44.2
datasets==3.0.1
accelerate==0.34.2
tokenizers==0.19.1

# ASR
openai-whisper==20231117

# Evaluation
jiwer==3.0.4
scikit-learn==1.5.2

# TTS
gTTS==2.5.3

# Visualization (optional, for evaluation reports)
matplotlib==3.9.2
seaborn==0.13.2

# Config
PyYAML==6.0.2

# Utilities
numpy==1.26.4
scipy==1.14.1

# Testing
pytest==8.3.3
httpx==0.27.2

# Logging
python-json-logger==2.0.7